+++ extensions/amp-apester-media/0.1/amp-apester-media.js
@@ -13,7 +13,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import {CONSENT_POLICY_STATE} from '../../../src/consent-state';
 import {CSS} from '../../../build/amp-apester-media-0.1.css';
 import {CustomEventReporterBuilder} from '../../../src/extension-analytics';
 import {IntersectionObserverApi} from '../../../src/intersection-observer-polyfill';
@@ -24,18 +23,14 @@ import {dict} from '../../../src/utils/object';
 import {
   extractTags,
   generatePixelURL,
-  getPageUrl,
   getPlatform,
   registerEvent,
   setFullscreenOff,
   setFullscreenOn,
 } from './utils';
-import {
-  getConsentPolicyInfo,
-  getConsentPolicyState,
-} from '../../../src/consent';
 import {getLengthNumeral, isLayoutSizeDefined} from '../../../src/layout';
-import {handleCompanion} from './monetization/monetization';
+import {handleCompanionDisplay} from './monetization/companion/display';
+import {handleCompanionVideo} from './monetization/companion/video';
 import {removeElement} from '../../../src/dom';
 import {setStyles} from '../../../src/style';
 
@@ -314,134 +309,6 @@ class AmpApesterMedia extends AMP.BaseElement {
     }
   }
 
-  /**
-   * @return {!JsonObject}
-   * */
-  getConsent_() {
-    const consentStatePromise = getConsentPolicyState(this.element).catch(
-      err => {
-        dev().error(TAG, 'Error determining consent state', err);
-        return CONSENT_POLICY_STATE.UNKNOWN;
-      }
-    );
-    const consentStringPromise = getConsentPolicyInfo(this.getAmpDoc()).catch(
-      err => {
-        dev().error(TAG, 'Error determining consent string', err);
-        return null;
-      }
-    );
-    return Promise.all([consentStatePromise, consentStringPromise]);
-  }
-  //todo test right behavior
-  /**
-   * @param {CONSENT_POLICY_STATE} consentPolicyState
-   * @return {number, number, string}
-   */
-  getConsentStateValue_(consentPolicyState) {
-    switch (consentPolicyState) {
-      case CONSENT_POLICY_STATE.SUFFICIENT:
-        return {gdpr: 1, user_consent: 1};
-      case CONSENT_POLICY_STATE.INSUFFICIENT:
-      case CONSENT_POLICY_STATE.UNKNOWN:
-        return {gdpr: 1, user_consent: 0};
-      case CONSENT_POLICY_STATE.UNKNOWN_NOT_REQUIRED:
-      default:
-        return {gdpr: 0, user_consent: 1};
-    }
-  }
-
-  /**
-   * @param {!JsonObject} interactionModel
-   * @param {?string} campaignId
-   * @return {!JsonObject}
-   */
-  getSrMacros_(interactionModel, campaignId) {
-    return this.getConsent_().then(consentRes => {
-      const {interactionId, publisherId, publisher} = interactionModel;
-      const macros = Object.assign(
-        {
-          param1: interactionId,
-          param2: publisherId,
-          param6: campaignId,
-          page_url: getPageUrl(),
-        },
-        this.getConsentStateValue_(consentRes[0])
-      );
-      const gdprString = consentRes[1];
-      if (gdprString) {
-        macros.param4 = gdprString;
-      }
-      if (publisher && publisher.groupId) {
-        macros.param7 = `apester.com,${publisher.groupId}`;
-        macros.schain = `1.0,1!apester.com,${publisher.groupId},1,,,,`;
-      }
-      return macros;
-    });
-  }
-
-  /**
-   * @return {!JsonObject}
-   */
-  getCompanionVideoAdSize_() {
-    const adWidth = this.element.clientWidth;
-    const adRatio = 0.6;
-    const adHeight = Math.ceil(adWidth * adRatio);
-    return {width: adWidth, height: adHeight};
-  }
-
-  /**
-   * @param {!JsonObject} companionRawSettings
-   * @return {!JsonObject}
-   */
-  extractCompanionSrSettings_(companionRawSettings) {
-    if (
-      !companionRawSettings ||
-      !companionRawSettings.video ||
-      !companionRawSettings.video.enabled ||
-      !companionRawSettings.video.provider === 'sr'
-    ) {
-      return null;
-    }
-    const res = {};
-    const {video} = companionRawSettings;
-    res.videoTag = video.videoTag;
-    if (video.companion.enabled) {
-      res.location = 'companionAbove';
-    } else if (video.companion_below.enabled) {
-      res.location = 'companionBelow';
-    } else {
-      return null;
-    }
-    res.size = this.getCompanionVideoAdSize_();
-    return res;
-  }
-
-  /**
-   * @param {JsonObject} companionSettings
-   * @param {JsonObject} media
-   * @param {string} campaignId
-   * @return {!Element}
-   */
-  constructCompanionSr_(companionSettings, media, campaignId) {
-    if (!companionSettings) {
-      return null;
-    }
-    const {videoTag, size} = companionSettings || {};
-    const ampAd = this.element.ownerDocument.createElement('amp-ad');
-    ampAd.setAttribute('type', 'blade');
-    return this.getSrMacros_(media, campaignId).then(macros => {
-      ampAd.setAttribute('data-blade_player_type', 'bladex');
-      ampAd.setAttribute('servingDomain', 'ssr.streamrail.net');
-      ampAd.setAttribute('width', size.width);
-      ampAd.setAttribute('height', size.height);
-      ampAd.setAttribute('data-blade_macros', JSON.stringify(macros));
-      ampAd.setAttribute('data-blade_player_id', videoTag);
-      ampAd.setAttribute('data-blade_api_key', '5857d2ee263dc90002000001');
-      ampAd.classList.add('amp-apester-companion');
-      return ampAd;
-    });
-  }
-
   /** @override */
   layoutCallback() {
     this.element.classList.add('amp-apester-container');
@@ -458,37 +325,6 @@ class AmpApesterMedia extends AMP.BaseElement {
         const media = /** @type {JsonObject} */ (this.embedOptions_.playlist
           ? payload[Math.floor(Math.random() * payload.length)]
           : payload);
-        const monetizationSettings = media['campaignData'] || {};
-        const companionRawSettings = monetizationSettings['companionOptions'];
-        const companionCampaignOptions =
-          monetizationSettings['companionCampaignOptions'] || {};
-        // const companionDisplaySettings = this.extractCompanionDisplaySettings_(
-        //   companionRawSettings
-        // );
-
-        // const companionDisplayElement = this.constructCompanionDisplayAd_(
-        //   companionDisplaySettings
-        // );
-        const {companionCampaignId} = companionCampaignOptions;
-        const companionSrSettings = this.extractCompanionSrSettings_(
-          companionRawSettings
-        );
-        if (companionSrSettings) {
-          this.constructCompanionSr_(
-            companionSrSettings,
-            media,
-            companionCampaignId
-          ).then(companionVideoSrElement => {
-            const companionSrElement =
-              companionSrSettings.location === 'companionBelow'
-                ? this.element.nextSibling
-                : this.element;
-            this.element.parentNode.insertBefore(
-              companionVideoSrElement,
-              companionSrElement
-            );
-          });
-        }
         const interactionId = media['interactionId'];
         const usePlayer = media['usePlayer'];
         const src = this.constructUrlFromMedia_(interactionId, usePlayer);
@@ -508,13 +344,8 @@ class AmpApesterMedia extends AMP.BaseElement {
             this.element.appendChild(overflow);
             this.element.appendChild(iframe);
 
-            handleCompanion(media['campaignData'], this.element);
-            // if (companionDisplayElement) {
-            //   this.element.parentNode.insertBefore(
-            //     companionDisplayElement,
-            //     this.element.nextSibling
-            //   );
-            // }
+            handleCompanionDisplay(media['campaignData'], this.element);
+            handleCompanionVideo(media, this.element);
           })
           .then(() => {
             return this.loadPromise(iframe).then(() => {
